refactor: move non-macro methods out
